+++ ambient_download/src/weather/download.rs
@@ -49,14 +49,14 @@ pub fn get_weather(
 
     let empty_str = String::new();
     let date = subcmd_args
-        .get_one::<String>("end-dates")
+        .get_one::<String>("start-date")
         .unwrap_or(&empty_str);
     let num_days = *subcmd_args.get_one::<u16>("days").unwrap_or(&1);
 
     let dl_date = if date.is_empty() {
-        yesterday()
+        end_of_day(&yesterday())
     } else {
-        let parse_date = format!("{date} 00:00:30 {}", config.tz_offset);
+        let parse_date = format!("{date} 23:59:30 {}", config.tz_offset);
 
         let Ok(parsed_date) = DateTime::parse_from_str(&parse_date, "%F %T %:z") else {
             let err_msg = format!("Could not parse {parse_date}.");
@@ -66,37 +66,9 @@ pub fn get_weather(
         parsed_date.with_timezone(&Local)
     };
 
-    download_weather(
-        &end_of_day(&dl_date)?,
-        num_days,
-        &creds,
-        &config,
-        detail_level,
-    )?;
+    download_weather(&dl_date, num_days, &creds, &config, detail_level)?;
 
     Ok(())
-}
-
-/// Get yesterday's weather information.
-#[allow(clippy::module_name_repetitions)]
-pub fn download_weather_data(
-    date: &DateTime<Local>,
-    creds: &Query,
-) -> Result<String, Box<dyn Error>> {
-    let limit = creds.limit.unwrap_or(288); // 288 is the maximum number of records that can be downloaded.
-    let end_date_str = date.to_rfc3339();
-
-    let Some(mac_address) = &creds.mac_address else {
-        return Err("MAC address not provided.".into());
-    };
-
-    let url = format!(
-        "https://rt.ambientweather.net/v1/devices/{mac_address}?apiKey={}&applicationKey={}&endDate={end_date_str}&limit={limit}",
-        creds.api_key, creds.app_key
-    );
-    let resp = reqwest::blocking::get(url)?.text()?;
-
-    Ok(resp)
 }
 
 /// Write the information to a JSON file
@@ -138,27 +110,73 @@ pub fn filename_from_datetime(date: &DateTime<Local>, ext: &str) -> String {
     filename
 }
 
-/// Download the weather information for a number of days, ending with the date specified
+/// Download the weather information for a number of days, starting with the date specified
 fn download_weather(
-    end_date: &DateTime<Local>,
+    start_date: &DateTime<Local>,
     num_days: u16,
     creds: &Query,
     config: &config::Config,
     detail_level: DetailLevel,
 ) -> Result<(), Box<dyn Error>> {
     if detail_level > DetailLevel::Quiet {
-        log::info!("Getting weather information for {num_days} days ending at {end_date}.");
+        log::info!("Getting weather information for {num_days} days starting at {start_date}.");
     }
 
     if num_days > 365 * 3 {
         return Err("The maximum number of days is 1095 (3 years).".into());
     }
 
-    let mut date = *end_date;
+    let mut date = *start_date;
     let output_folder = &config.output_folder;
 
+    let limit = creds.limit.unwrap_or(288); // 288 is the maximum number of records that can be downloaded.
+
+    // Create a reqwest blocking client that will be used to download the weather information.
+    let timeout = std::time::Duration::from_secs(&config.sleep_time + 5);
+    let kas = if config.sleep_time > 20 { 10 } else { 5 };
+    let keepalive = std::time::Duration::from_secs(kas);
+    let client = reqwest::blocking::Client::builder()
+        .timeout(Some(timeout))
+        .tcp_keepalive(keepalive)
+        .user_agent("ambient-downloader")
+        .build()?;
+
+    if detail_level > DetailLevel::Quiet {
+        log::info!(
+            "Sleeping for {} seconds between each download to avoid rate limiting.",
+            config.sleep_time
+        );
+    }
+
     for d in 0..num_days {
-        let weather_info = download_weather_data(&date, creds)?;
+        let Some(mac_address) = &creds.mac_address else {
+            return Err("MAC address not provided.".into());
+        };
+
+        let url = format!(
+            "https://rt.ambientweather.net/v1/devices/{mac_address}?apiKey={}&applicationKey={}&endDate={}&limit={limit}",
+            creds.api_key, creds.app_key, date.to_rfc3339()
+        );
+
+        let res = client.get(url).send();
+        let weather_info;
+        match res {
+            Ok(resp) => {
+                if resp.status().is_success() {
+                    weather_info = resp.text()?;
+                } else {
+                    let err_msg = format!(
+                        "Error downloading weather information. Status code: {}.",
+                        resp.status().as_str()
+                    );
+                    return Err(err_msg.into());
+                }
+            }
+            Err(e) => {
+                let err_msg = format!("Error downloading weather information: {e}.");
+                return Err(err_msg.into());
+            }
+        }
 
         let output_file_name = filename_from_datetime(&date, "json");
         let full_path = format!("{output_folder}/{output_file_name}");
@@ -168,15 +186,9 @@ fn download_weather(
             log::info!("Wrote {bytes_written} bytes to {full_path}.");
         };
 
-        date -= chrono::Duration::days(1);
+        date += chrono::Duration::days(1);
 
         if num_days > 1 && d < num_days - 1 {
-            if detail_level > DetailLevel::Quiet {
-                log::info!(
-                    "Sleeping for {} seconds to avoid rate limiting.",
-                    config.sleep_time
-                );
-            }
             sleep(std::time::Duration::from_secs(config.sleep_time));
         }
     }
@@ -190,11 +202,8 @@ pub fn yesterday() -> DateTime<Local> {
 }
 
 /// Return a date at the end of the day/
-pub fn end_of_day(date: &DateTime<Local>) -> Result<DateTime<Local>, Box<dyn Error>> {
+pub fn end_of_day(date: &DateTime<Local>) -> DateTime<Local> {
     // Find the end of the day
-    let eod_local = date
-        .with_time(NaiveTime::from_hms_opt(23, 59, 30).unwrap_or_default())
-        .unwrap();
-
-    Ok(eod_local)
+    date.with_time(NaiveTime::from_hms_opt(23, 59, 30).unwrap_or_default())
+        .unwrap()
 }